Update README.md
In Listing 1.1, lines 000100 and 000200 are the IDENTIFICATION DIVISION. This division is used to
identify basic information about the program. In this example, the IDENTIFICATION DIVISION contains
only the PROGRAM-ID, HELLO.
Line 000300 is the ENVIRONMENT DIVISION, which is used to identify the environment in which the
program is running. Remember that COBOL is intended to run on many different types of machines, and this
section is used to handle the differences between various computers. In this case, the program has no
specialized machine requirements, so the ENVIRONMENT DIVISION is empty.
Line 000400 is the DATA DIVISION, which will contain any data that the program operates on. This
program has no data, so the DATA DIVISION is empty.
Lines 000500 through 001100 are the PROCEDURE DIVISION. This is the meat of the program--the part
that does the work intended by the programmer. The PROCEDURE DIVISION contains two paragraphs at line
000700 (PROGRAM-BEGIN) and line 001000 (PROGRAM-DONE). The term paragraph has a special
definition in COBOL that will be covered a bit later in today's lesson. All the actual work in this program is
done by line 000800.